requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Supabase
supabase==2.3.0
postgrest-py==0.13.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
asyncpg==0.29.0

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0

# AI/ML (OpenRouter for ESP simulation & chatbot)
httpx==0.26.0
openai==1.10.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
